# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VyOS API",
    "description": "REST proxy for one or more VyOS router devices. Exposes CRUD operations over interfaces, VRFs, VLANs, firewall policies, and firewall address groups. Each device is identified by the name supplied in the VYOS_HOSTS environment variable.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8082",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/devices/{device_id}/vrfs": {
      "parameters": [
        { "$ref": "#/components/parameters/device_id" }
      ],
      "get": {
        "tags": ["vrfs"],
        "summary": "List all VRFs",
        "operationId": "listVRFs",
        "responses": {
          "200": {
            "description": "VRF list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/VRFInfo" }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/DeviceNotFound" },
          "422": { "$ref": "#/components/responses/DeviceRejected" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      },
      "post": {
        "tags": ["vrfs"],
        "summary": "Create a VRF",
        "operationId": "createVRF",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateVRFRequest" },
              "example": { "name": "MGMT", "table": "100", "description": "management-vrf" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "VRF created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VRFInfo" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/DeviceNotFound" },
          "422": { "$ref": "#/components/responses/DeviceRejected" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      }
    },
    "/devices/{device_id}/vrfs/{vrf}": {
      "parameters": [
        { "$ref": "#/components/parameters/device_id" },
        { "$ref": "#/components/parameters/vrf" }
      ],
      "get": {
        "tags": ["vrfs"],
        "summary": "Get a VRF",
        "operationId": "getVRF",
        "responses": {
          "200": {
            "description": "VRF details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VRFInfo" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      },
      "put": {
        "tags": ["vrfs"],
        "summary": "Update a VRF",
        "description": "Updates one or more fields. Omit fields that should not change.",
        "operationId": "updateVRF",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateVRFRequest" },
              "example": { "description": "updated-description" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "VRF updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VRFInfo" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/DeviceNotFound" },
          "422": { "$ref": "#/components/responses/DeviceRejected" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      },
      "delete": {
        "tags": ["vrfs"],
        "summary": "Delete a VRF",
        "operationId": "deleteVRF",
        "responses": {
          "204": { "description": "VRF deleted" },
          "404": { "$ref": "#/components/responses/DeviceNotFound" },
          "422": { "$ref": "#/components/responses/DeviceRejected" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      }
    },
    "/devices/{device_id}/vlans": {
      "parameters": [
        { "$ref": "#/components/parameters/device_id" }
      ],
      "get": {
        "tags": ["vlans"],
        "summary": "List all VLANs",
        "description": "Iterates all interface types looking for `vif` (802.1Q) subinterfaces.",
        "operationId": "listVLANs",
        "responses": {
          "200": {
            "description": "VLAN list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/VLANInfo" }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/DeviceNotFound" },
          "422": { "$ref": "#/components/responses/DeviceRejected" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      },
      "post": {
        "tags": ["vlans"],
        "summary": "Create a VLAN subinterface",
        "description": "Creates a `vif` subinterface under the specified parent interface. `address` is optional.",
        "operationId": "createVLAN",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateVLANRequest" },
              "example": {
                "interface": "eth0",
                "type": "ethernet",
                "vlan_id": 100,
                "address": "10.100.0.1/24",
                "description": "servers-vlan"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "VLAN created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VLANInfo" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/DeviceNotFound" },
          "422": { "$ref": "#/components/responses/DeviceRejected" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      }
    },
    "/devices/{device_id}/vlans/{interface}/{vlan_id}": {
      "parameters": [
        { "$ref": "#/components/parameters/device_id" },
        { "$ref": "#/components/parameters/interface" },
        { "$ref": "#/components/parameters/vlan_id" },
        { "$ref": "#/components/parameters/iface_type_query" }
      ],
      "get": {
        "tags": ["vlans"],
        "summary": "Get a VLAN subinterface",
        "operationId": "getVLAN",
        "responses": {
          "200": {
            "description": "VLAN details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VLANInfo" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      },
      "put": {
        "tags": ["vlans"],
        "summary": "Update a VLAN subinterface",
        "description": "Replaces the address and/or description on the subinterface. If `address` is provided all existing addresses are replaced. `type` defaults to `ethernet`.",
        "operationId": "updateVLAN",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateVLANRequest" },
              "example": { "type": "ethernet", "address": "10.100.0.2/24" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "VLAN updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VLANInfo" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/DeviceNotFound" },
          "422": { "$ref": "#/components/responses/DeviceRejected" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      },
      "delete": {
        "tags": ["vlans"],
        "summary": "Delete a VLAN subinterface",
        "description": "Removes the entire `vif` subinterface config node. Defaults to `type=ethernet`.",
        "operationId": "deleteVLAN",
        "responses": {
          "204": { "description": "VLAN deleted" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/DeviceNotFound" },
          "422": { "$ref": "#/components/responses/DeviceRejected" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "device_id": {
        "name": "device_id",
        "in": "path",
        "required": true,
        "description": "Device name as registered in VYOS_HOSTS",
        "schema": { "type": "string", "example": "router1" }
      },
      "interface": {
        "name": "interface",
        "in": "path",
        "required": true,
        "description": "Interface name (e.g. `eth0`, `bond0`, `lo`)",
        "schema": { "type": "string", "example": "eth0" }
      },
      "vlan_id": {
        "name": "vlan_id",
        "in": "path",
        "required": true,
        "description": "802.1Q VLAN ID (1–4094)",
        "schema": { "type": "integer", "minimum": 1, "maximum": 4094, "example": 100 }
      },
      "vrf": {
        "name": "vrf",
        "in": "path",
        "required": true,
        "description": "VRF name",
        "schema": { "type": "string", "example": "MGMT" }
      },
...
      "iface_type_query": {
        "name": "type",
        "in": "query",
        "required": false,
        "description": "Interface type in the VyOS config tree. Defaults to `ethernet`.",
        "schema": {
          "type": "string",
          "default": "ethernet",
          "enum": ["ethernet", "loopback", "bonding", "bridge", "wireguard", "dummy"]
        }
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "example": "device not found: router99" }
        }
      },
...
      "VRFInfo": {
        "type": "object",
        "required": ["name", "table"],
        "properties": {
          "name":        { "type": "string", "example": "MGMT" },
          "table":       { "type": "string", "description": "Linux routing table ID", "example": "100" },
          "description": { "type": "string", "example": "management-vrf" }
        }
      },
      "CreateVRFRequest": {
        "type": "object",
        "required": ["name", "table"],
        "properties": {
          "name":        { "type": "string", "example": "MGMT" },
          "table":       { "type": "string", "description": "Linux routing table ID (unique across VRFs)", "example": "100" },
          "description": { "type": "string", "example": "management-vrf" }
        }
      },
      "UpdateVRFRequest": {
        "type": "object",
        "properties": {
          "table":       { "type": "string", "example": "101" },
          "description": { "type": "string", "example": "updated-description" }
        }
      },
      "VLANInfo": {
        "type": "object",
        "required": ["interface", "type", "vlan_id", "addresses"],
        "properties": {
          "interface":   { "type": "string", "example": "eth0" },
          "type":        { "type": "string", "example": "ethernet" },
          "vlan_id":     { "type": "integer", "minimum": 1, "maximum": 4094, "example": 100 },
          "addresses":   { "type": "array", "items": { "type": "string" }, "example": ["10.100.0.1/24"] },
          "description": { "type": "string", "example": "servers-vlan" }
        }
      },
      "CreateVLANRequest": {
        "type": "object",
        "required": ["interface", "type", "vlan_id"],
        "properties": {
          "interface":   { "type": "string", "example": "eth0" },
          "type":        { "type": "string", "example": "ethernet" },
          "vlan_id":     { "type": "integer", "minimum": 1, "maximum": 4094, "example": 100 },
          "address":     { "type": "string", "description": "Optional IPv4 address in CIDR notation", "example": "10.100.0.1/24" },
          "description": { "type": "string", "example": "servers-vlan" }
        }
      },
      "UpdateVLANRequest": {
        "type": "object",
        "properties": {
          "type":        { "type": "string", "description": "Interface type; defaults to `ethernet`", "example": "ethernet" },
          "address":     { "type": "string", "description": "Replacement IPv4 address in CIDR notation", "example": "10.100.0.2/24" },
          "description": { "type": "string", "example": "updated-vlan" }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Missing or invalid request fields",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "interface, type, and address are required" }
          }
        }
      },
      "NotFound": {
        "description": "Device or resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "interface not found" }
          }
        }
      },
      "DeviceNotFound": {
        "description": "The requested device_id is not registered",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "device not found: router99" }
          }
        }
      },
      "DeviceRejected": {
        "description": "The VyOS device rejected the operation (invalid config, constraint violation, etc.)",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "device rejected operation: Configuration path already exists" }
          }
        }
      },
      "DeviceError": {
        "description": "Could not communicate with the VyOS device (network error, TLS failure, timeout)",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "device communication error: dial tcp 192.168.1.1:443: connection refused" }
          }
        }
      }
    }
  }
}